feat(api): use quality dimensions enum for metadata validation
