+++ workers/ohsome_quality_analyst/api/request_models.py
@@ -21,7 +21,6 @@ from ohsome_quality_analyst.definitions import (
     get_fid_fields,
     get_indicator_names,
     get_project_keys,
-    get_quality_dimension_keys,
     get_report_names,
     get_topic_keys,
 )
@@ -32,9 +31,6 @@ from ohsome_quality_analyst.utils.validators import validate_geojson
 IndicatorEnum = Enum("IndicatorEnum", {name: name for name in get_indicator_names()})
 ReportEnum = Enum("ReportEnum", {name: name for name in get_report_names()})
 TopicEnum = Enum("TopicEnum", {name: name for name in get_topic_keys()})
-QualityDimensionEnum = Enum(
-    "QualityDimensionEnum", {key: key for key in get_quality_dimension_keys()}
-)
 DatasetEnum = Enum("DatasetNames", {name: name for name in get_dataset_names()})
 FidFieldEnum = Enum("FidFieldEnum", {name: name for name in get_fid_fields()})
 ProjectEnum = Enum("ProjectEnum", {name: name for name in get_project_keys()})
