+++ workers/ohsome_quality_analyst/api/response_models.py
@@ -3,12 +3,12 @@ from pydantic import BaseModel, validator
 from ohsome_quality_analyst import __version__
 from ohsome_quality_analyst.api.request_models import (
     IndicatorEnum,
-    QualityDimensionEnum,
     ReportEnum,
     TopicEnum,
 )
 from ohsome_quality_analyst.definitions import ATTRIBUTION_URL
 from ohsome_quality_analyst.indicators.models import IndicatorMetadata
+from ohsome_quality_analyst.quality_dimensions.definitions import QualityDimensionEnum
 from ohsome_quality_analyst.quality_dimensions.models import QualityDimension
 from ohsome_quality_analyst.reports.models import ReportMetadata
 from ohsome_quality_analyst.topics.models import TopicDefinition
