+++ workers/ohsome_quality_analyst/quality_dimensions/definitions.py
@@ -1,0 +1,49 @@
+"""Global Variables and Functions."""
+import os
+from enum import Enum
+
+import yaml
+
+from ohsome_quality_analyst.quality_dimensions.models import QualityDimension
+from ohsome_quality_analyst.utils.helper import get_module_dir
+
+
+def load_quality_dimensions() -> dict[str, QualityDimension]:
+    """Read definitions of quality dimensions.
+
+    Returns:
+        A dict with all quality dimensions included.
+    """
+    directory = get_module_dir("ohsome_quality_analyst.quality_dimensions")
+    file = os.path.join(directory, "quality_dimensions.yaml")
+    with open(file, "r") as f:
+        raw = yaml.safe_load(f)
+    quality_dimensions = {}
+    for k, v in raw.items():
+        quality_dimensions[k] = QualityDimension(**v)
+    return quality_dimensions
+
+
+def get_quality_dimensions() -> dict[str, QualityDimension]:
+    quality_dimensions = load_quality_dimensions()
+    return quality_dimensions
+
+
+def get_quality_dimension(qd_key: str) -> QualityDimension:
+    quality_dimensions = get_quality_dimensions()
+    try:
+        return quality_dimensions[qd_key]
+    except KeyError as error:
+        raise KeyError(
+            "Invalid quality dimension0key. Valid quality dimension keys are: "
+            + str(quality_dimensions.keys())
+        ) from error
+
+
+def get_quality_dimension_keys() -> list[str]:
+    return [str(t) for t in load_quality_dimensions().keys()]
+
+
+QualityDimensionEnum = Enum(
+    "QualityDimensionEnum", {key: key for key in get_quality_dimension_keys()}
+)
